+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2025-10-20
+## \[1.1.0] - 2025-12-29
+
 ### Added
-- Initial public release of SPDX Report
+
+* SCA-57445: Addition of Notices Text as ExtractedText in the SBOM report
+
+## \[1.0.0] - 2025-10-20
+
+### Added
+
+* Initial public release of SPDX Report